+++ CHANGELOG.rst
@@ -6,12 +6,17 @@
   - The ``euphonic-dos``, ``euphonic-dispersion`` and
     ``euphonic-intensity-map`` command-line tools can now read
     files that don't contain eigenvectors, if eigenvectors are
-    not required for the chosen options
+    not required for the chosen options.
 
 - Changes:
 
   - ``euphonic.cli.force_constants_from_file`` and ``modes_from_file``
-    have been deprecated in favour of ``euphonic.cli.load_data_from_file``
+    have been deprecated in favour of ``euphonic.cli.load_data_from_file``.
+  - Using ``Spectrum1D/1DCollection/2D.broaden`` on an axis with unequal
+    bin widths is now deprecated, as broadening is performed via convolution,
+    which is incorrect in this case. In the future, this will raise a
+    ``ValueError``. To broaden anyway, ``method='convolve'`` can be supplied,
+    which will just emit a warning.
 
 `v0.6.3 <https://github.com/pace-neutrons/Euphonic/compare/v0.6.2...v0.6.3>`_
 ------